Moved modules dealing with exact TN circuit representation to s separate folder.
