+++ pytket/extensions/cutensornet/exact/__init__.py
@@ -15,17 +15,11 @@
 
 from pytket.extensions.cutensornet.exact.utils import circuit_statevector_postselect
 
-from pytket.extensions.cutensornet.exact.tensor_network_convert import (
+from .tensor_network_convert import (
     TensorNetwork,
     PauliOperatorTensorNetwork,
     ExpectationValueTensorNetwork,
     tk_to_tensor_network,
     measure_qubits_state,
 )
-
-from .backends import CuTensorNetBackend
-
-# _metadata.py is copied to the folder after installation.
-from ._metadata import __extension_version__, __extension_name__  # type: ignore
-
